requirements.txt
pillow
pandas
faker
pytesseract
jiwer
google-cloud-vision
